+++ src/didl.h
@@ -33,7 +33,7 @@ didl_add_tag (struct buffer_s *out, char *tag, char *value);
 void
 didl_add_param (struct buffer_s *out, char *param, char *value);
 void
-didl_add_value (struct buffer_s *out, char *param, off_t value);
+didl_add_value (struct buffer_s *out, char *param, uint32_t value);
 void
 didl_add_short_item (buffer_t *out,
     uint32_t id, dlna_item_t *item, uint32_t containerid);
